rename user actions file bck to types, keys back to types
