+++ app/src/main/report/actionTypes/UsersActionTypes.ts
@@ -1,18 +1,18 @@
 import {User} from "../../shared/models/Generated";
 
-export enum UserActionKeys {
+export enum UserActionTypes {
     REPORT_READERS_FETCHED = "REPORT_READERS_FETCHED",
     REPORT_READER_REMOVED = "REPORT_READER_REMOVED",
     REPORT_READER_ADDED = "REPORT_READER_ADDED"
 }
 
 export interface ReportReadersFetched {
-    type: UserActionKeys.REPORT_READERS_FETCHED;
+    type: UserActionTypes.REPORT_READERS_FETCHED;
     data: User[];
 }
 
 export interface ReportReaderRemoved {
-    type: UserActionKeys.REPORT_READER_REMOVED;
+    type: UserActionTypes.REPORT_READER_REMOVED;
     data: string;
 }
