+++ app/src/main/report/actions/userActionCreators.ts
@@ -1,6 +1,6 @@
 import {Dispatch} from "redux";
 import {GlobalState} from "../../shared/reducers/GlobalState";
-import {ReportReaderRemoved, ReportReadersFetched, UserActionKeys} from "../actionTypes/UsersActions";
+import {ReportReaderRemoved, ReportReadersFetched, UserActionTypes} from "../actionTypes/UsersActionTypes";
 import {UserService} from "../services/UserService";
 import {User} from "../../shared/models/Generated";
 
@@ -10,7 +10,7 @@ export const userActionCreators = {
         return async (dispatch: Dispatch<any>, getState: () => GlobalState) => {
             const users: User[] = await (new UserService(dispatch, getState)).getReportReaders(reportName);
             dispatch({
-                type: UserActionKeys.REPORT_READERS_FETCHED,
+                type: UserActionTypes.REPORT_READERS_FETCHED,
                 data: users
             } as ReportReadersFetched);
         }
@@ -22,7 +22,7 @@ export const userActionCreators = {
 
             if (result == "OK") {
                 dispatch({
-                    type: UserActionKeys.REPORT_READER_REMOVED,
+                    type: UserActionTypes.REPORT_READER_REMOVED,
                     data: username
                 } as ReportReaderRemoved);
             }
