+++ app/src/main/report/reducers/userReducer.ts
@@ -1,5 +1,5 @@
 import {User} from "../../shared/models/Generated";
-import {UserAction, UserActionKeys} from "../actionTypes/UsersActions";
+import {UserAction, UserActionTypes} from "../actionTypes/UsersActionTypes";
 
 export interface UsersState {
     reportReaders: User[];
@@ -11,9 +11,9 @@ export const usersInitialState: UsersState = {
 
 export const usersReducer = (state = usersInitialState, action: UserAction): UsersState => {
     switch (action.type) {
-        case UserActionKeys.REPORT_READERS_FETCHED:
+        case UserActionTypes.REPORT_READERS_FETCHED:
             return {...state, reportReaders: action.data};
-        case UserActionKeys.REPORT_READER_REMOVED:
+        case UserActionTypes.REPORT_READER_REMOVED:
             return {...state, reportReaders: state.reportReaders.filter(u => u.username != action.data)};
         default:
             return state;
